requirements.txt
# Core Deep Learning
tensorflow>=2.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Tracking
tqdm>=4.65.0

# Grad-CAM Visualization
opencv-python>=4.8.0
